port to XNG
This commit prepares the port to XNG.

This adds an abstraction layer that enables drop-in implementations for the
network interface layer. The code generation now only has to specify the
contents of the main function, the structs for the hypervisor implementation and
the network interface layer.

The static global variables are not sync anymore so users of the code generation
can use no_std. It is safe to use them in an unsafe context, if this is done
only during initialization and the contents are then cloned / copied to use them
in a multi-threaded context. (This is what the generated code does.)

The network-partition-echo crate is merged with the the network-partition-linux
crate, since the echo example is only configured and has no own code, besides
the `include` of the generated code. This has the disadvantage of having to run
the `build.rs` even though the library from network-partition-linux is only to
be used in another crate and the library does not require the code-generation.
The disadvantage of having to maintain two crates instead of one justifies
this small performance overhead. The generated code is not compiled if only the
library is used.

See also dadada/zynq-network-partition for the deployment on the Zynq boards and
dadada/network-partition-xng for the XNG-port.