Add a builder and macros for configuration
This adds a builder object and macros that makes creating a
configuration more accessible. The fields holding the configuration
contents are marked as private.

This change is motivated by the fact that becomes tedious to carry
around the const generics when building configurations in the
[config-generator](https://gitlab.dlr.de/projekt-resilienz/demonstrator)

This also prevents interior mutability of the configuration, which
should make it easier to change the content of the configuration
without requiring as many code changes and prevents accidental
modification in other modules.

One tradeoff is that now every virtual link gets as large as the largest
virtual link. But instead the size of every individual virtual link does
not need to be specified.