router-partition-macros: init
Add a macro crate for generating a router partition declaratively. The
declarative configuration is provided as macro-attributes and
describes both the router's and the partition's static configurations.

In terms of macro-hygiene one current limitation is that the nested
macros involved require the user to explicitly add dependencies to
`a653rs` and `network-partition`, with the `macros` features enabled.